HW_1
Assignment for week one. It takes about half a minute to 1 minute to
load and filter the data, so if you check the code in R you might prefer
to skip first couple of lines of plot2.r, plot3.r, and plot4.r unless
you're ok waiting a bit. Thank you